Some tweaks based on some changes in Consistent
Added scope.$.fire to invoke event handlers while search through the scope parents for them and invoking with this set to the scope.

Changed Consistent to call its own change/keyup functions first so it has updated the scope by the time our user-defined functions run. This matches the expectation.